requirements/base.txt
Django==2.2
mysqlclient==1.4.6
Pillow==5.2.0
PyMySQL==0.9.3
pypng==0.0.20
PyQRCode==1.2.1
pytz==2020.1
qrcode==6.1
six==1.15.0
sqlparse==0.3.1
